requirements.txt
langchain
langchain_core
langchain_community
langchain-google-genai
python-dotenv